Not uploading images. Saving them to local disk with FileSystem helper methods.

+++ TrapEase/Network/EUCNetwork.m
@@ -147,7 +147,7 @@ static NSString * baseUrl = @"http://trap.euclidsoftware.com";
     EUCNetwork * network = [EUCNetwork sharedNetwork];
     
     
-    [network.sessionManager POST:[NSString stringWithFormat:@"/%@/%ld", resource, numIds]
+    [network.sessionManager POST:[NSString stringWithFormat:@"/%@/%ld", resource, (long)numIds]
                       parameters:nil
                          success:^(NSURLSessionDataTask *task, id responseObject) {
                              NSDictionary * result = (NSDictionary *) responseObject;
@@ -168,7 +168,7 @@ static NSString * baseUrl = @"http://trap.euclidsoftware.com";
     EUCNetwork * network = [EUCNetwork sharedNetwork];
     
     
-    [network.sessionManager PUT:[NSString stringWithFormat:@"/%@/%ld", resource, resourceId]
+    [network.sessionManager PUT:[NSString stringWithFormat:@"/%@/%ld", resource, (long)resourceId]
                      parameters:params
                         success:successBlock
                         failure:failureBlock];
@@ -183,7 +183,7 @@ static NSString * baseUrl = @"http://trap.euclidsoftware.com";
     
     NSMutableURLRequest *request = [[AFHTTPRequestSerializer serializer]
                                     multipartFormRequestWithMethod:@"POST"
-                                    URLString:[NSString stringWithFormat:@"%@/file/%@/%ld", baseUrl, resource, resourceId]
+                                    URLString:[NSString stringWithFormat:@"%@/file/%@/%ld", baseUrl, resource, (long)resourceId]
                                     parameters:nil
                                     constructingBodyWithBlock:^(id<AFMultipartFormData> formData) {
                                         [formData appendPartWithFileData:data
@@ -201,7 +201,7 @@ static NSString * baseUrl = @"http://trap.euclidsoftware.com";
     NSURLSessionUploadTask *uploadTask = [network.sessionManager uploadTaskWithStreamedRequest:request
                                                                                       progress:&progress
                                                                              completionHandler:^(NSURLResponse *response, id responseObject, NSError *error) {
-                                                                                 NSLog(@"Resource: %@, resourceId; %ld", resource, resourceId);
+                                                                                 NSLog(@"Resource: %@, resourceId; %ld", resource, (long)resourceId);
                                                                                  if (error) {
                                                                                      NSLog(@"Error: %@", error);
                                                                                  } else {
